added switch between game modes issue #80
