+++ src/modules/assemblyGame/AssemblyGame.jsx
@@ -1,0 +1,12 @@
+/**
+ * @fileoverview Provides Assembly game for CSTEM's Educational Games.
+ * @author James Isaacks <james.isaacks@roguedevstudios.com>
+ * */
+
+ import React from "react"
+
+ const AssemblyGame = (level) => {
+     return <div className = "assemblyGame">This is game is not ready yet</div>
+ }
+ 
+ export default AssemblyGame
