--- src/modules/assemblyGame/assemblygame.js
@@ -1,5 +1,0 @@
-/**
- * @fileoverview Provides Assembly game for CSTEM's Educational Games.
- * @author James Isaacks <james.isaacks@roguedevstudios.com>
- * */
-
