--- src/modules/huntGame/huntGame.js
@@ -1,156 +1,0 @@
-/**
- * @fileoverview Provides Item Hunt game for CSTEM's Educational Games.
- * @author James Isaacks <james.isaacks@roguedevstudios.com>
- * @author Kevin Segarra
- * */
-
-function huntGame() {
-    <div className="huntGame">
-    </div>
-}
-
-// TODO: Remove this section after refactor.
-/*var body = "" +
-"<section class=\"draggable-items\">" +
-"</section>" +
-"<section class=\"matching-pairs\">" +
-"</section>" +
-"";
-
-document.querySelector('body').innerHTML += (body);
-
-const brands = [
-  {
-    itemLink: "",
-    itemName: ""
-  },
-  {
-    itemLink: "",
-    itemName: ""
-  },
-  {
-    itemLink: "",
-    itemName: ""
-  },
-  {
-    itemLink: "",
-    itemName: ""
-  },
-  {
-    itemLink: "",
-    itemName: ""
-  }
-];
-let correct = 0;
-let total = 0;
-const totalDraggableItems = 5;
-const totalMatchingPairs = 5; // Should be <= totalDraggableItems
-
-const draggableItems = document.querySelector(".draggable-items");
-const matchingPairs = document.querySelector(".matching-pairs");
-let draggableElements;
-let droppableElements;
-
-initiateGame();
-
-function initiateGame() {
-  const randomDraggableItems = generateRandomItemsArray(totalDraggableItems, brands);
-  const randomDroppableBrands = totalMatchingPairs<totalDraggableItems ? generateRandomItemsArray(totalMatchingPairs, randomDraggableItems) : randomDraggableItems;
-  const alphabeticallySortedRandomDroppableItem = [...randomDroppableBrands].sort((a,b) => a.itemName.toLowerCase().localeCompare(b.itemName.toLowerCase()));
-
-  // Create "draggable-items" and append to DOM
-  for(let i=0; i<randomDraggableItems.length; i++) {
-    draggableItems.insertAdjacentHTML("beforeend", `
-      <img src="${randomDraggableItems[i].itemLink}" class="draggable" draggable="true" id="${randomDraggableItems[i].itemLink}">
-    `);
-  }
-
-  // Create "matching-pairs" and append to DOM
-  for(let i=0; i<alphabeticallySortedRandomDroppableItem.length; i++) {
-    matchingPairs.insertAdjacentHTML("beforeend", `
-      <div class="matching-pair">
-        <span class="label">${alphabeticallySortedRandomDroppableItem[i].itemName}</span>
-        <span class="droppable" data-brand="${alphabeticallySortedRandomDroppableItem[i].itemLink}"></span>
-      </div>
-    `);
-  }
-
-  draggableElements = document.querySelectorAll(".draggable");
-  droppableElements = document.querySelectorAll(".droppable");
-
-  draggableElements.forEach(elem => {
-    elem.addEventListener("dragstart", dragStart);
-    // elem.addEventListener("drag", drag);
-    // elem.addEventListener("dragend", dragEnd);
-  });
-
-  droppableElements.forEach(elem => {
-    elem.addEventListener("dragenter", dragEnter);
-    elem.addEventListener("dragover", dragOver);
-    elem.addEventListener("dragleave", dragLeave);
-    elem.addEventListener("drop", drop);
-  });
-}
-
-// Drag and Drop Functions
-
-//Events fired on the drag target
-
-function dragStart(event) {
-  event.dataTransfer.setData("text", event.target.id); // or "text/plain"
-}
-
-//Events fired on the drop target
-
-function dragEnter(event) {
-  if(event.target.classList && event.target.classList.contains("droppable") && !event.target.classList.contains("dropped")) {
-    event.target.classList.add("droppable-hover");
-  }
-}
-
-function dragOver(event) {
-  if(event.target.classList && event.target.classList.contains("droppable") && !event.target.classList.contains("dropped")) {
-    event.preventDefault();
-  }
-}
-
-function dragLeave(event) {
-  if(event.target.classList && event.target.classList.contains("droppable") && !event.target.classList.contains("dropped")) {
-    event.target.classList.remove("droppable-hover");
-  }
-}
-
-function drop(event) {
-  event.preventDefault();
-  event.target.classList.remove("droppable-hover");
-  const draggableElementBrand = event.dataTransfer.getData("text");
-  const droppableElementBrand = event.target.getAttribute("data-brand");
-  const isCorrectMatching = draggableElementBrand===droppableElementBrand;
-  total++;
-  if(isCorrectMatching) {
-    const draggableElement = document.getElementById(draggableElementBrand);
-    event.target.classList.add("dropped");
-    draggableElement.classList.add("dragged");
-    draggableElement.setAttribute("draggable", "false");
-    correct++;
-  }
-  setTimeout(() => {
-  }, 200);
-  if(correct===Math.min(totalMatchingPairs, totalDraggableItems)) { // Game Over!!
-    setTimeout(() => {
-    }, 200);
-  }
-}
-
-// Auxiliary functions
-function generateRandomItemsArray(n, originalArray) {
-  let res = [];
-  let clonedArray = [...originalArray];
-  if(n>clonedArray.length) n=clonedArray.length;
-  for(let i=1; i<=n; i++) {
-    const randomIndex = Math.floor(Math.random()*clonedArray.length);
-    res.push(clonedArray[randomIndex]);
-    clonedArray.splice(randomIndex, 1);
-  }
-  return res;
-}*/
